refactor(server): move the fifteen inline test modules to sibling files (#1918)
Fourth crate of the per-crate migration campaign. Thirteen files are a
straight #[path] sibling move; handlers/search/pipeline.rs carried two
separately named inline modules (parse_fusion_strategy_tests and
timeout_response_tests) and was split by hand into one sibling each,
keeping both module names. The sorted test-name list is byte-identical
before and after (157 lib tests).

config.rs fell under the 1000-line budget once its 556-line test module
moved out, so its now-meaningless entry leaves file-budgets-baseline.txt;
the inline-tests baseline shrinks by the fifteen velesdb-server lines
(2278 lines paid), leaving 158 files.
