+++ scripts/inline-tests-baseline.txt
@@ -123,21 +123,6 @@ crates/velesdb-python/src/options.rs	194
 crates/velesdb-python/src/streaming_config.rs	22
 crates/velesdb-python/src/utils.rs	222
 crates/velesdb-python/src/velesql.rs	147
-crates/velesdb-server/src/auth.rs	115
-crates/velesdb-server/src/config.rs	559
-crates/velesdb-server/src/handlers/admin.rs	88
-crates/velesdb-server/src/handlers/graph/mod.rs	306
-crates/velesdb-server/src/handlers/graph/stream.rs	51
-crates/velesdb-server/src/handlers/graph/types.rs	59
-crates/velesdb-server/src/handlers/helpers.rs	68
-crates/velesdb-server/src/handlers/match_query.rs	141
-crates/velesdb-server/src/handlers/metrics.rs	79
-crates/velesdb-server/src/handlers/points/mod.rs	25
-crates/velesdb-server/src/handlers/query/mod.rs	75
-crates/velesdb-server/src/handlers/search/pipeline.rs	163
-crates/velesdb-server/src/lib.rs	461
-crates/velesdb-server/src/rate_limit.rs	48
-crates/velesdb-server/src/tls.rs	40
 crates/velesdb-wasm/src/agent.rs	58
 crates/velesdb-wasm/src/fusion.rs	338
 crates/velesdb-wasm/src/graph_persistence.rs	5
